support for SVG, PNG and gnuplot>=4.6
for gnuplot>=4.6 tics and axis color will be set, as it would be
invisible else

you can use -o file.svg to create an SVG file instead of PNG.
This is better for zooming or detailed analysis. If you want to display
many tiny plots on a single screen it might be better to use PNG with
has smaller size.